+++ lib/src/reel_text.dart
@@ -1,6 +1,5 @@
 import 'dart:async';
 import 'dart:math' as math;
-import 'dart:ui' as ui;
 
 import 'package:flutter/material.dart';
 
@@ -10,4 +9,5 @@ part 'reel_text_controller.dart';
 part 'reel_text_glyphs.dart';
 part 'reel_text_metrics.dart';
 part 'reel_text_plan.dart';
+part 'reel_text_selection.dart';
 part 'reel_text_widget.dart';